Fixup some things
