+++ enterprise/cmd/executor/internal/run/run.go
@@ -20,6 +20,10 @@ import (
 )
 
 func RunRun(cliCtx *cli.Context, logger log.Logger, cfg *config.Config) error {
+	return StandaloneRunRun(cliCtx.Context, logger, cfg, cliCtx.Bool("verify"))
+}
+
+func StandaloneRunRun(ctx context.Context, logger log.Logger, cfg *config.Config, runVerifyChecks bool) error {
 	if err := cfg.Validate(); err != nil {
 		return err
 	}
@@ -42,14 +46,14 @@ func RunRun(cliCtx *cli.Context, logger log.Logger, cfg *config.Config) error {
 	opts := apiWorkerOptions(cfg, queueTelemetryOptions)
 
 	// TODO: This is too similar to the RunValidate func. Make it share even more code.
-	if cliCtx.Bool("verify") {
+	if runVerifyChecks {
 		// Then, validate all tools that are required are installed.
 		if err := validateToolsRequired(cfg.UseFirecracker); err != nil {
 			return err
 		}
 
 		// Validate git is of the right version.
-		if err := validateGitVersion(cliCtx.Context); err != nil {
+		if err := validateGitVersion(ctx); err != nil {
 			return err
 		}
 
@@ -60,13 +64,13 @@ func RunRun(cliCtx *cli.Context, logger log.Logger, cfg *config.Config) error {
 		if err != nil {
 			return err
 		}
-		if err := validateSrcCLIVersion(cliCtx.Context, client, opts.QueueOptions.BaseClientOptions.EndpointOptions); err != nil {
+		if err := validateSrcCLIVersion(ctx, client, opts.QueueOptions.BaseClientOptions.EndpointOptions); err != nil {
 			return err
 		}
 
 		if cfg.UseFirecracker {
 			// Validate ignite is installed.
-			if err := validateIgniteInstalled(cliCtx.Context); err != nil {
+			if err := validateIgniteInstalled(ctx); err != nil {
 				return err
 			}
 
@@ -81,7 +85,7 @@ func RunRun(cliCtx *cli.Context, logger log.Logger, cfg *config.Config) error {
 	}
 
 	nameSet := janitor.NewNameSet()
-	ctx, cancel := context.WithCancel(cliCtx.Context)
+	ctx, cancel := context.WithCancel(ctx)
 	worker, err := worker.NewWorker(observationCtx, nameSet, opts)
 	if err != nil {
 		cancel()
