+++ enterprise/cmd/executor/singlebinary/doc.go
@@ -1,0 +1,3 @@
+// singlebinary is the entrypoint for the sourcegraph single binary process. It must
+// not be used or imported from anywhere else.
+package singlebinary
