+++ enterprise/dev/ci/integration/executors/run.sh
@@ -27,7 +27,7 @@ trap cleanup EXIT
 export POSTGRES_IMAGE="us.gcr.io/sourcegraph-dev/postgres-12-alpine:${CANDIDATE_VERSION}"
 export SERVER_IMAGE="us.gcr.io/sourcegraph-dev/server:${CANDIDATE_VERSION}"
 export EXECUTOR_IMAGE="us.gcr.io/sourcegraph-dev/executor:${CANDIDATE_VERSION}"
-export EXECUTOR_FRONTEND_PASSWORD=none
+export EXECUTOR_FRONTEND_PASSWORD="hunter2hunter2hunter2"
 export SOURCEGRAPH_LICENSE_GENERATION_KEY="${SOURCEGRAPH_LICENSE_GENERATION_KEY:-""}"
 export TMP_DIR
 export DATA
