+++ sg.config.yaml
@@ -99,8 +99,8 @@ env:
 
   # Required for frontend and executor to communicate
   EXECUTOR_FRONTEND_URL: http://localhost:3080
-  EXECUTOR_FRONTEND_PASSWORD: none
-  EXECUTOR_QUEUE_DISABLE_ACCESS_TOKEN_INSECURE: true
+  # Must match the secret defined in the site config.
+  EXECUTOR_FRONTEND_PASSWORD: hunter2hunter2hunter2
 
   # Disable firecracker inside executor in dev
   EXECUTOR_USE_FIRECRACKER: false
